Update more int->NSInteger for window numbers
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40348 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSDisplayServer.h
@@ -93,8 +93,8 @@ APPKIT_EXPORT NSString *GSScreenNumber;
 - (BOOL) slideImage: (NSImage*)image from: (NSPoint)from to: (NSPoint)to;
 - (void) restrictWindow: (NSInteger)win toImage: (NSImage*)image;
 - (int) findWindowAt: (NSPoint)screenLocation 
-           windowRef: (int*)windowRef 
-           excluding: (int)win;
+           windowRef: (NSInteger*)windowRef
+           excluding: (NSInteger)win;
 
 
 /* Screen information */
